Move `reserved_email` and `user` into new `account` package.
